pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kroget"
version = "0.1.0"
description = "Kroger shopping CLI"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Kroget Contributors" }]
license = { text = "MIT" }
keywords = ["kroger", "cli"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
  "typer[all]",
  "rich",
  "httpx",
  "pydantic>=2",
  "python-dotenv",
  "textual",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "ruff",
]

[project.scripts]
kroget = "kroget.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["kroget*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
markers = [
  "integration: marks tests requiring live Kroger creds",
]
